Add more documentation.
Move cleanup.ts to cli.ts in anticipation of a more general CLI tool for faast.js. Restructure the cli to take a command as the first argument (the only valid command is `cleanup` for now). Add package.json `bin` configuration to deploy this as the `faastjs` command line program.

Add graph of cost analyzer results using 200 repetitions of 100M random numbers for each memory size on AWS.

Limit cost analyzer concurrency of configurations to 8. This helps ensure some configurations finish before too many others are started, which is more satisfying to observe interactively than waiting for completion of a long test.