Make documentation more github pages friendly.
The file "/index.html" is a target of links generated by the api-documenter for the "Home" breadcrumb navigation at the top of each page. The api-documenter blows away its output directory, so the "doc" script target needs to copy it back from etc/ (which is just a convenient place to put it) back into doc/api.