refactor(serialization): Factor serialization code into one place.
Instead of ad-hoc calls to JSON.stringify and JSON.parse, we now factor serialization and
deserialization into a single place so it can be changed centrally. One side effect is that throttle
caches are checked during serialization of arguments and values. This extra validation caught some
issues with the way caching worked with google pricing, which was also refactored to be safer.

re #5