test: Add verbose log URL output to limits test.
Adding the log URL to a test enables easier debugging of issues in the testsuite because the test's
logs will be available on the cloud provider for at least a full day. This change makes it easy to
turn on remote detailed logging for any faast instance. We turn it on for the timeout limit test to
tryto gain more detail when this particularly fragile test fails in the future.