improvement(webpack): complain when webpack issues errors or warnings
It's too easy to fall into the trap of including a library (potentially deep in the dependency
chain) that webpack cannot pack properly. We have packageJson for this scenario, but in practice
it's hard to realize this is the solution to the problem of a broken invocation. We make webpack
failures more explicit by issuing a warning if there are webpack errors or warnings, and provide a
suggestion to turn on webpack debugging as well as a link to the docs for packageJson. This should
make it easier to debug failed invocations.