feat: allow Error objects to be passed as args and return values
In some cases error objects are returned not as thrown errors but as informational objects. This
happens often if there are many sub-tasks each of which might fail, but do not preclude making
progress with other sub-tasks. Enabling returning Error objects makes it more convenient to write
such code. Limitation: custom subclasses of Error will result in serialization errors.