fix(aws): use Lambda Destinations to route errors more reliably on AWS
Lambda Destinations allows for much faster routing of errors in case a function terminates without
sending a response explicitly. This fixes a number of corner cases, such as getting timeout and
memory exceeded errors for functions not using a child process. Also, this allows us to remove the
provision for dead letter queue messages from the provider interface, since AWS was the only user of
this in the past. Also some minor refactorings of the Pump interface to make arguments more
explicit.